requirements.txt
streamlit
google-generativeai
base64